requirements.txt
openfeature_sdk==0.8.3
splitio_client[cpphash,asyncio]==10.5.1